requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
aiohttp=3.12.14=py312h8a5da7c_0
aiosignal=1.4.0=pyhd8ed1ab_0
annotated-types=0.7.0=pyhd8ed1ab_1
anyio=4.9.0=pyh29332c3_0
async-timeout=4.0.3=pyhd8ed1ab_0
attrs=25.3.0=pyh71513ae_0
brotli-python=1.1.0=py312h2ec8cdc_3
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.5=hb9d3cd8_0
ca-certificates=2025.7.14=hbd8a1cb_0
cachetools=5.5.2=pyhd8ed1ab_0
certifi=2025.7.14=pyhd8ed1ab_0
cffi=1.17.1=py312h06ac9bb_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.1.8=py312h06a4308_0
colorama=0.4.6=pyhd8ed1ab_1
cryptography=45.0.5=py312hda17c39_0
distro=1.9.0=pyhd8ed1ab_1
dnspython=2.4.2=py312h06a4308_0
email-validator=2.2.0=pyhd8ed1ab_1
email_validator=2.2.0=hd8ed1ab_1
exceptiongroup=1.3.0=pyhd8ed1ab_0
fastapi=0.116.1=pyhe01879c_0
fastapi-cli=0.0.8=pyhd8ed1ab_0
filetype=1.2.0=pyhd8ed1ab_0
frozenlist=1.7.0=py312h447239a_0
google-ai-generativelanguage=0.6.15=pyhd8ed1ab_0
google-api-core=2.25.1=pyhd8ed1ab_0
google-api-python-client=2.176.0=pyhff2d567_0
google-auth=2.40.3=pyhd8ed1ab_0
google-auth-httplib2=0.2.0=pyhd8ed1ab_1
google-generativeai=0.8.5=pyh29332c3_0
googleapis-common-protos=1.70.0=pyhd8ed1ab_0
greenlet=3.2.3=py312h2ec8cdc_0
grpcio=1.71.0=py312hdcb7bd4_1
h11=0.16.0=pyhd8ed1ab_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httplib2=0.22.0=pyhd8ed1ab_1
httptools=0.6.4=py312h5eee18b_0
httpx=0.28.1=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_1
jinja2=3.1.6=py312h06a4308_0
jiter=0.10.0=py312h12e396e_0
jsonpatch=1.33=pyhd8ed1ab_1
jsonpointer=3.0.0=py312h7900ff3_1
langchain=0.3.26=pymin312_hff2d567_0
langchain-core=0.3.68=pyhd8ed1ab_0
langchain-google-genai=2.0.10=pyhd8ed1ab_0
langchain-openai=0.3.27=pyhd8ed1ab_0
langchain-text-splitters=0.3.8=pyhd8ed1ab_0
langsmith=0.3.45=pyhd8ed1ab_0
ld_impl_linux-64=2.44=h1423503_1
libabseil=20250127.1=cxx17_hbbce691_0
libblas=3.9.0=32_h59b9bed_openblas
libcblas=3.9.0=32_he106b2a_openblas
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgfortran=15.1.0=h69a702a_3
libgfortran5=15.1.0=hcea5267_3
libgomp=15.1.0=h767d61c_3
libgrpc=1.71.0=h8e591d7_1
liblapack=3.9.0=32_h7ac8fdf_openblas
liblzma=5.8.1=hb9d3cd8_2
libnsl=2.0.1=hb9d3cd8_1
libopenblas=0.3.30=pthreads_h94d23a6_0
libprotobuf=5.29.3=h501fc15_1
libre2-11=2025.06.26=hba17884_0
libsqlite=3.50.2=hee844dc_2
libstdcxx=15.1.0=h8f9b012_3
libstdcxx-ng=15.1.0=h4852527_3
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=h5eee18b_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
markdown-it-py=2.2.0=py312h06a4308_1
markupsafe=3.0.2=py312h5eee18b_0
mdurl=0.1.0=py312h06a4308_0
multidict=6.6.3=py312h178313f_0
ncurses=6.5=h2d0b736_3
numpy=1.26.4=py312heda63a1_0
openai=1.95.1=pyhd8ed1ab_0
openssl=3.5.1=h7b32b05_0
orjson=3.10.18=py312h680f630_1
packaging=24.2=pyhd8ed1ab_2
pip=25.1.1=pyh8b19718_0
propcache=0.3.1=py312h178313f_0
proto-plus=1.26.1=pyhd8ed1ab_0
protobuf=5.29.3=py312h0f4f066_0
pyasn1=0.6.1=pyhd8ed1ab_2
pyasn1-modules=0.4.2=pyhd8ed1ab_0
pycparser=2.22=pyh29332c3_1
pydantic=2.11.7=pyh3cfb1c2_0
pydantic-core=2.33.2=py312h680f630_0
pygments=2.19.1=py312h06a4308_0
pyopenssl=25.1.0=pyhd8ed1ab_0
pyparsing=3.2.3=pyhd8ed1ab_1
pysocks=1.7.1=pyha55dd90_7
python=3.12.11=h9e4cc4f_0_cpython
python-dotenv=1.1.1=pyhe01879c_0
python-multipart=0.0.20=py312h06a4308_0
python_abi=3.12=7_cp312
pyu2f=0.1.5=pyhd8ed1ab_1
pyyaml=6.0.2=py312h178313f_2
re2=2025.06.26=h9925aae_0
readline=8.2=h8c095d6_2
regex=2024.11.6=py312h66e93f0_0
requests=2.32.4=pyhd8ed1ab_0
requests-toolbelt=1.0.0=pyhd8ed1ab_1
rich=13.9.4=py312h06a4308_0
rich-toolkit=0.14.8=pyhe01879c_0
rsa=4.9.1=pyhd8ed1ab_0
setuptools=80.9.0=pyhff2d567_0
shellingham=1.5.0=py312h06a4308_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pyhd8ed1ab_1
sqlalchemy=2.0.41=py312h66e93f0_0
starlette=0.46.2=py312h06a4308_0
tenacity=9.1.2=pyhd8ed1ab_0
tiktoken=0.9.0=py312h14ff09d_0
tk=8.6.13=noxft_hd72426e_102
tqdm=4.67.1=pyhd8ed1ab_1
typer=0.16.0=pyh167b9f4_0
typer-slim=0.16.0=pyhe01879c_0
typer-slim-standard=0.16.0=hf964461_0
typing-extensions=4.14.1=h4440ef1_0
typing-inspection=0.4.1=pyhd8ed1ab_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h78e105d_0
uritemplate=4.2.0=pyhd8ed1ab_0
urllib3=2.5.0=pyhd8ed1ab_0
uvicorn=0.35.0=py312h06a4308_0
uvicorn-standard=0.35.0=py312h06a4308_0
uvloop=0.21.0=py312h5eee18b_0
watchfiles=0.24.0=py312h4aa5aa6_2
websockets=15.0.1=py312h5eee18b_0
wheel=0.45.1=pyhd8ed1ab_1
yaml=0.2.5=h7f98852_2
yarl=1.20.1=py312h178313f_0
zstandard=0.23.0=py312h66e93f0_2
